# docs/openapi.yaml
openapi: 3.1.0
info:
  title: klaxond API
  version: 0.14.26
  summary: Notification bridge backend API
  description: |
    klaxond receives monitoring webhooks, renders notification payloads, manages
    routing/inhibition policy, and exposes a zero-build admin UI.

    Browser sessions use the `klaxond_session` cookie plus `X-Klaxond-CSRF` on
    same-origin mutations. Scoped machine clients should use Bearer tokens.
    Sensitive browser mutations can also return `428 Precondition Required`
    with `X-Klaxond-Reauth: required`; call `/api/auth/reauth` or complete passkey
    reauthentication before retrying.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
  /openapi.yaml:
    get:
      tags: [Public]
      operationId: getOpenApiYaml
      summary: Download the OpenAPI document.
      security: []
      responses:
        "200":
          description: YAML OpenAPI document.
          content:
            application/yaml:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/openapi.yaml:
    get:
      tags: [Public]
      operationId: getOpenApiYamlAlias
      summary: Download the OpenAPI document from the API namespace.
      security: []
      responses:
        "200":
          description: YAML OpenAPI document.
          content:
            application/yaml:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /swagger:
    get:
      tags: [Public]
      operationId: swaggerUiLegacy
      summary: Render the self-hosted Swagger UI from the legacy root alias.
      deprecated: true
      security: []
      responses:
        "200":
          description: Swagger UI HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/docs:
    get:
      tags: [Public]
      operationId: swaggerUiDocs
      summary: Render the self-hosted Swagger UI from the API documentation namespace.
      security: []
      responses:
        "200":
          description: Swagger UI HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/swagger:
    get:
      tags: [Public]
      operationId: swaggerUi
      summary: Render the self-hosted Swagger UI from the API namespace.
      security: []
      responses:
        "200":
          description: Swagger UI HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/swagger-ui:
    get:
      tags: [Public]
      operationId: swaggerUiStandalone
      summary: Render the self-hosted Swagger UI from the explicit Swagger UI alias.
      security: []
      responses:
        "200":
          description: Swagger UI HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /legal:
    get:
      tags: [Public]
      operationId: legalIndex
      summary: Redirect to the default public legal page.
      security: []
      responses:
        "302":
          description: Redirects to `/legal/privacy`.
        "4XX":
          $ref: "#/components/responses/ClientError"
  /legal/privacy:
    get:
      tags: [Public]
      operationId: legalPrivacy
      summary: Render the public privacy notice page.
      security: []
      responses:
        "200":
          description: Privacy notice HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/methods:
    get:
      tags: [Auth]
      operationId: listAuthMethods
      summary: Return available authentication methods.
      description: Returns stable auth method names and whether each method is enabled in the active Klaxond runtime configuration.
      security: []
      responses:
        "200":
          description: Authentication method availability.
          content:
            application/json:
              schema:
                type: object
                required: [methods]
                properties:
                  methods:
                    type: array
                    items:
                      type: object
                      required: [method, enabled]
                      properties:
                        method:
                          type: string
                          enum: [password, oidc, totp, passkey, hardware_key, trusted_proxy, ldap, api_token, magic_link]
                        enabled:
                          type: boolean
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/password-policy:
    get:
      tags: [Auth]
      operationId: getPasswordPolicy
      summary: Return the active local password policy.
      description: Returns the shared Argon2id password policy limits that the frontend uses for local password fields.
      security: []
      responses:
        "200":
          description: Password policy limits.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordPolicy"
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: currentUser
      summary: Return the current user, auth mode, scopes and browser CSRF token.
      security:
        - bearerAuth: []
        - basicAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/passkey/login:
    get:
      tags: [Auth]
      operationId: passkeyLoginPage
      summary: Render the public passkey login page.
      security: []
      responses:
        "200":
          description: Passkey login HTML.
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/step-up:
    get:
      tags: [Auth]
      operationId: stepUpPage
      summary: Render the public MFA / step-up page.
      description: The page is reachable before the app session is issued and is guarded by the one-time step-up token.
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
        - in: query
          name: return_to
          schema:
            type: string
            default: /status
      responses:
        "200":
          description: MFA / step-up HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/status:
    get:
      tags: [Auth]
      operationId: getStepUpStatus
      summary: Return the pending MFA / step-up challenge metadata.
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pending challenge metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepUpStatus"
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/status:
    get:
      tags: [Operations]
      operationId: getStatus
      summary: Return runtime status, channel health and recent activity.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Status payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPayload"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/setup-status:
    get:
      tags: [Operations]
      operationId: getSetupStatus
      summary: Return readiness/setup checks used by the UI.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Setup checks grouped by area.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/channel-test-matrix:
    get:
      tags: [Operations]
      operationId: getChannelTestMatrix
      summary: Dry-run channel connectivity matrix; sends no notification.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Channel matrix.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/logs:
    get:
      tags: [Operations]
      operationId: searchLogs
      summary: Search backend/frontend log entries.
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Level"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated log entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLogs"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/audit:
    get:
      tags: [Operations]
      operationId: searchAudit
      summary: Search security and configuration audit entries.
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated audit entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAudit"
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/config:
    get:
      tags: [Auth]
      operationId: getAuthConfig
      summary: Return redacted auth settings and available auth modes/scopes.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Redacted auth settings.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Auth]
      operationId: updateAuthConfig
      summary: Replace auth settings after validation.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthConfigUpdate"
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/tokens:
    get:
      tags: [Auth]
      operationId: listTokens
      summary: List API keys and PATs without token secrets.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Token metadata and available scopes.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Auth]
      operationId: createToken
      summary: Create a scoped API key or PAT.
      description: The raw token is returned once; store the prefix for later identification.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "200":
          description: Created token and one-time secret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/auth/passkey/credentials:
    get:
      tags: [Auth]
      operationId: listPasskeys
      summary: List registered passkeys and public WebAuthn settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Passkey metadata.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/config/backup:
    get:
      tags: [Config]
      operationId: downloadConfigBackup
      summary: Download the current TOML configuration.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: "`klaxond.toml` attachment."
          content:
            application/toml:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/config/export:
    get:
      tags: [Config]
      operationId: exportConfigBundle
      summary: Export a complete settings bundle.
      description: Admin-only export containing TOML, sidecar JSON files, auth sidecars and runtime-derived secret placeholders.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Full settings bundle JSON attachment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigBundle"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/config/backups:
    get:
      tags: [Config]
      operationId: listConfigBackups
      summary: List automatic config backups created before restore.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Backup metadata.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/channel-config:
    get:
      tags: [Routing]
      operationId: getChannelConfig
      summary: Return channel configuration and redacted secret presence/source flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Channel configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateChannelConfig
      summary: Update channel fields, including TOML-backed secrets when env overrides are not used.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/ntfy-topics:
    get:
      tags: [Routing]
      operationId: getNtfyTopics
      summary: Return ntfy topics, known severities and token presence flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: ntfy topic configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateNtfyTopics
      summary: Replace ntfy topic sidecar configuration.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/ingest-auth:
    get:
      tags: [Routing]
      operationId: getIngestAuth
      summary: Return webhook ingest authentication settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Ingest auth settings.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateIngestAuth
      summary: Update webhook ingest authentication settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/cascade-config:
    get:
      tags: [Routing]
      operationId: getCascadeConfig
      summary: Return cascade tier list and runtime/default flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Cascade configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateCascadeConfig
      summary: Update cascade tiers and defaults.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/dedup-config:
    get:
      tags: [Policy]
      operationId: getDedupConfig
      summary: Return dedup settings and pending queue counts.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Dedup configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Policy]
      operationId: updateDedupConfig
      summary: Replace dedup settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /inhibitions:
    get:
      tags: [Policy]
      operationId: listInhibitionsLegacyAlias
      summary: Legacy alias for `/api/inhibitions`.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Active inhibitions.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/inhibitions:
    get:
      tags: [Policy]
      operationId: listInhibitions
      summary: Return active in-memory suppressions with TTL.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Active inhibitions.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/inhibition-rules:
    get:
      tags: [Policy]
      operationId: listInhibitionRules
      summary: Return configured inhibition rules.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Inhibition rule configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Policy]
      operationId: updateInhibitionRules
      summary: Replace inhibition rules.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/schedules:
    get:
      tags: [Policy]
      operationId: listSchedules
      summary: Return scheduler state.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Scheduler status.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Policy]
      operationId: updateSchedules
      summary: Replace schedule configuration.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/acks:
    get:
      tags: [Policy]
      operationId: listAcks
      summary: Return runtime acknowledgement/snooze state.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Acknowledgement state snapshot.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: klaxond scoped token
      description: API keys and PATs created from `/api/auth/tokens`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: klaxond_session
      description: Signed browser session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-Klaxond-CSRF
      description: Required on browser-originated mutations when using a session cookie.
    basicAuth:
      type: http
      scheme: basic
      description: Username/password auth for non-browser clients in basic or LDAP mode.
  parameters:
...
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
...
    ClientError:
      description: Generic client-side error, including invalid input, missing credentials or authorization failure depending on the active auth mode.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
...
    PasswordPolicy:
      type: object
      required: [min_length, max_length]
      properties:
        min_length:
          type: integer
          minimum: 1
          description: Minimum allowed local password length.
        max_length:
          type: integer
          minimum: 1
          description: Maximum accepted local password length.
...
    SetupStatus:
      type: object
      additionalProperties: true
      properties:
        checks:
          type: array
          items:
            type: object
...
security:
  - bearerAuth: []
  - sessionCookie: []